+++ examples/torchrec_dlrm/README.md
@@ -5,6 +5,14 @@ This example shows how to serve a Deep Learning Recommendation Model (DLRM) with
 The DLRM is an open source model for personalization and recommendation use cases published by Meta. More information can be found in this [paper](https://arxiv.org/abs/1906.00091) and this [blog post](https://ai.facebook.com/blog/dlrm-an-advanced-open-source-deep-learning-recommendation-model/).
 TorchRec is Meta's open source library for recommender systems in Pytorch. More information on TorchRec can be found in the [official docs](https://pytorch.org/torchrec/).
 
+# Install torchrec using the following command
+This needs pytorch 1.12 to work
+
+```bash
+pip install torchrec==0.2.0
+```
+
+
 --- **This example requires CUDA version >= 11.3**  ---
 
 In this example we will first create and archive the DLRM into a mar file which is subsequently registered in a TorchServe instance. Finally, we run inference using curl.
