+++ examples/torchrec_dlrm/dlrm_model_config.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import List
+
+
+@dataclass
+class DLRMModelConfig:
+    dense_arch_layer_sizes: List[int]
+    dense_in_features: int
+    embedding_dim: int
+    id_list_features_keys: List[str]
+    num_embeddings_per_feature: List[int]
+    over_arch_layer_sizes: List[int]
